pyproject.toml
[tool.poetry]
name = "MlopsBasics"
version = "0.1.0"
description = "Self study of mlops based on graviraja/MLOps-Basics repo."
authors = ["gatheluck"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
torch = "^1.9.0"
torchvision = "^0.10.0"
hydra-core = "^1.1.0"
pytorch-lightning = "^1.4.0"
comet-ml = "^3.13.2"
mlflow = "^1.19.0"
albumentations = "^1.0.3"
transformers = "^4.10.2"
datasets = "^1.12.1"
scikit-learn = "^0.24.2"
hydra-colorlog = "^1.1.0"
wandb = "^0.12.2"
pandas = "^1.3.3"

[tool.poetry.dev-dependencies]
mypy = "^0.910"
black = "^21.7b0"
isort = "^5.9.2"
flake8 = "^3.9.2"
pytest-cov = "^2.12.1"

[tool.isort]
profile = "black"

[[tool.mypy.overrides]]
module = "pandas"
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"